fix(android): resolve build errors and missing imports
- Added missing asSharedFlow import in NativeEngine.kt
- Added missing zIndex import in MainActivity.kt
- Added missing hal/shared_memory_bridge.h include in ronin_jni.cpp
- Fixed printf format warning in hydration_manager.cpp

+++ src/models/hydration_manager.cpp
@@ -62,7 +62,7 @@ bool HydrationManager::hydrate(const std::string& model_path) {
 
     // Phase 2 Logic: Threshold Guard (1GB)
     uint64_t available = getAvailableRAM();
-    LOGI(TAG, "Hydration: Available RAM is %llu MB", available / (1024 * 1024));
+    LOGI(TAG, "Hydration: Available RAM is %llu MB", (unsigned long long)(available / (1024 * 1024)));
 
     if (available > 1024ULL * 1024ULL * 1024ULL) {
         LOGI(TAG, "Hydration: Attempting mlock for ultra-low latency.");
